requirements.txt
Django==4.2.7
opencv-python==4.8.0.74
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.0.0
gunicorn==21.2.0
whitenoise==6.5.0
django-crispy-forms==2.0
crispy-tailwind==0.5.0
matplotlib==3.7.2
scikit-learn==1.3.0
beautifulsoup4==4.12.2
